Refactor utils code structure -  Add `utils/helper.py` - Change `ColorCycler` plot.py to helper.py
