+++ src/pycirclize/utils/__init__.py
@@ -4,14 +4,14 @@ from pycirclize.utils.dataset import (
     load_eukaryote_example_dataset,
     load_prokaryote_example_file,
 )
-from pycirclize.utils.plot import ColorCycler
+from pycirclize.utils.helper import ColorCycler
 from pycirclize.utils.tree import TreeUtil
 
 __all__ = [
+    "plot",
     "fetch_genbank_by_accid",
     "load_eukaryote_example_dataset",
     "load_prokaryote_example_file",
-    "plot",
     "ColorCycler",
     "TreeUtil",
 ]
