+++ src/pycirclize/utils/helper.py
@@ -1,0 +1,83 @@
+from __future__ import annotations
+
+import matplotlib as mpl
+import numpy as np
+from matplotlib.colors import Colormap, to_hex
+
+
+class ColorCycler:
+    """Color Cycler Class"""
+
+    counter = 0
+    cmap: Colormap = mpl.colormaps["tab10"]
+
+    def __new__(cls, n: int | None = None) -> str:
+        """Get hexcolor cyclically from cmap by counter or user specified number
+
+        `ColorCycler()` works same as `ColorCycler.get_color()`
+
+        Parameters
+        ----------
+        n : int | None, optional
+            Number for color cycle. If None, counter class variable is used.
+
+        Returns
+        -------
+        hexcolor : str
+            Cyclic hexcolor string
+        """
+        return cls.get_color(n)
+
+    @classmethod
+    def reset_cycle(cls) -> None:
+        """Reset cycle counter"""
+        cls.counter = 0
+
+    @classmethod
+    def set_cmap(cls, name: str) -> None:
+        """Set colormap (Default: `tab10`)"""
+        cls.cmap = mpl.colormaps[name]
+        cls.counter = 0
+
+    @classmethod
+    def get_color(cls, n: int | None = None) -> str:
+        """Get hexcolor cyclically from cmap by counter or user specified number
+
+        Parameters
+        ----------
+        n : int | None, optional
+            Number for color cycle. If None, counter class variable is used.
+
+        Returns
+        -------
+        hexcolor : str
+            Cyclic hexcolor string
+        """
+        if n is None:
+            n = cls.counter
+            cls.counter += 1
+        return to_hex(cls.cmap(n % cls.cmap.N), keep_alpha=True)
+
+    @classmethod
+    def get_color_list(cls, n: int | None = None) -> list[str]:
+        """Get hexcolor list of colormap
+
+        Parameters
+        ----------
+        n : int | None, optional
+            If n is None, all(=cmap.N) hexcolors are extracted from colormap.
+            If n is specified, hexcolors are extracted from n equally divided colormap.
+
+        Returns
+        -------
+        hexcolor_list : list[str]
+            Hexcolor list
+        """
+        if n is None:
+            cmap_idx_list = list(range(0, cls.cmap.N))
+        elif n > 0:
+            cmap_idx_list = [int(i) for i in np.linspace(0, cls.cmap.N, n)]
+        else:
+            raise ValueError(f"{n=} is invalid number (Must be 'n > 0').")
+
+        return [to_hex(cls.cmap(i), keep_alpha=True) for i in cmap_idx_list]
